[OpenMP] Add interface for creating stack traces on the device
Summary:
This patch adds an interface for pushing stack traces onto the target device.
This is done by surrounding each function call in the device with a push / pop
pair of function calls. The pushed value is an index to a global string that
contains the source location information for each function call.