[OpenMP] Add basic RTL implementation of device stack traces
Summary:
This patch adds the basic support for recieving stack traces from the device. A
new plugin function `__tgt_rtl_get_stack_trace` was added which will print a
stack trace if any were availible from the device. The stack trace is built
using the static data added by the compiler and an internal state updated by
the device. The state is stored in host-mapped memory so that the device can
update it and it will be visible on the device.